backend/requirements.txt
Flask
jsonify
uuid
requests
google-genai
python-dotenv
flask_cors